Dumping solution part 1

+++ src/com/org/gnos/db/model/TruckHourCost.java
@@ -1,12 +1,13 @@
 package com.org.gnos.db.model;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
 public class TruckHourCost extends FixedOpexCost{
 	public TruckHourCost(){
 		super();
 		this.id = -1;
-		this.costData = new LinkedHashMap<Integer, Float>();
+		this.costData = new LinkedHashMap<Integer, BigDecimal>();
 
 	}
